backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
python-arango==7.9.1
motor==3.3.2

# NLP and ML
spacy==3.7.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
networkx==3.2.1
scikit-learn==1.4.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# CLI
typer[all]==0.9.0
rich==13.7.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
